requirements.txt
flask==2.3.3
flask-cors==4.0.0
sentence-transformers==2.2.2
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
transformers>=4.21.0
fastapi
uvicorn[standard]
open_clip_torch
timm
